create stitch snd stitchtype modules
