+++ KnittingChartPreview/Assets/Scripts/Stitch.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: eb2db12b50514835ab3be5308c6ccbeb
+timeCreated: 1707864903
