+++ KnittingChartPreview/Assets/Scripts/StitchType.cs
@@ -1,0 +1,8 @@
+namespace YarnGenerator
+{
+    public enum StitchType
+    {
+        KnitStitch,
+        PurlStitch
+    }
+}
