+++ KnittingChartPreview/Assets/Scripts/StitchType.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 71e3ee1c7c1c47fc92e48be72ab88f7b
+timeCreated: 1707874426
